+++ pkg/provider/acme/provider.go
@@ -13,12 +13,13 @@ import (
 	"sync"
 	"time"
 
-	"github.com/go-acme/lego/v4/certificate"
-	"github.com/go-acme/lego/v4/challenge"
-	"github.com/go-acme/lego/v4/challenge/dns01"
-	"github.com/go-acme/lego/v4/lego"
-	"github.com/go-acme/lego/v4/providers/dns"
-	"github.com/go-acme/lego/v4/registration"
+	"github.com/go-acme/lego/v5/acme"
+	"github.com/go-acme/lego/v5/certificate"
+	"github.com/go-acme/lego/v5/challenge"
+	"github.com/go-acme/lego/v5/challenge/dns01"
+	"github.com/go-acme/lego/v5/lego"
+	"github.com/go-acme/lego/v5/providers/dns"
+	"github.com/go-acme/lego/v5/registration"
 	ptypes "github.com/traefik/paerser/types"
 	"github.com/traefik/traefik/v2/pkg/config/dynamic"
 	"github.com/traefik/traefik/v2/pkg/log"
@@ -49,7 +50,7 @@ type Configuration struct {
 
 // SetDefaults sets the default values.
 func (a *Configuration) SetDefaults() {
-	a.CAServer = lego.LEDirectoryProduction
+	a.CAServer = lego.DirectoryURLLetsEncrypt
 	a.Storage = "acme.json"
 	a.KeyType = "RSA4096"
 	a.CertificatesDuration = 3 * 30 * 24 // 90 Days
@@ -244,14 +245,14 @@ func (p *Provider) getClient() (*lego.Client, error) {
 		return p.client, nil
 	}
 
-	account, err := p.initAccount(ctx)
+	account, err := p.initAccount()
 	if err != nil {
 		return nil, err
 	}
 
 	logger.Debug("Building ACME client...")
 
-	caServer := lego.LEDirectoryProduction
+	caServer := lego.DirectoryURLLetsEncrypt
 	if len(p.CAServer) > 0 {
 		caServer = p.CAServer
 	}
@@ -259,7 +260,6 @@ func (p *Provider) getClient() (*lego.Client, error) {
 
 	config := lego.NewConfig(account)
 	config.CADirURL = caServer
-	config.Certificate.KeyType = GetKeyType(ctx, p.KeyType)
 	config.UserAgent = fmt.Sprintf("containous-traefik/%s", version.Version)
 
 	client, err := lego.NewClient(config)
@@ -274,7 +274,10 @@ func (p *Provider) getClient() (*lego.Client, error) {
 			return nil, errR
 		}
 
-		account.Registration = reg
+		account.Registration = &Resource{
+			Body: reg.Account,
+			URI:  reg.Location,
+		}
 	}
 
 	// Save the account once before all the certificates generation/storing
@@ -293,6 +296,12 @@ func (p *Provider) getClient() (*lego.Client, error) {
 	if p.DNSChallenge != nil && len(p.DNSChallenge.Provider) > 0 {
 		logger.Debugf("Using DNS Challenge provider: %s", p.DNSChallenge.Provider)
 
+		if len(p.DNSChallenge.Resolvers) > 0 {
+			opts := &dns01.Options{RecursiveNameservers: p.DNSChallenge.Resolvers}
+
+			dns01.SetDefaultClient(dns01.NewClient(opts))
+		}
+
 		var provider challenge.Provider
 		provider, err = dns.NewDNSChallengeProviderByName(p.DNSChallenge.Provider)
 		if err != nil {
@@ -300,8 +309,6 @@ func (p *Provider) getClient() (*lego.Client, error) {
 		}
 
 		err = client.Challenge.SetDNS01Provider(provider,
-			dns01.CondOption(len(p.DNSChallenge.Resolvers) > 0,
-				dns01.AddRecursiveNameservers(p.DNSChallenge.Resolvers)),
 			dns01.PropagationWait(time.Duration(p.DNSChallenge.DelayBeforeCheck), p.DNSChallenge.DisablePropagationCheck),
 		)
 		if err != nil {
@@ -331,24 +338,19 @@ func (p *Provider) getClient() (*lego.Client, error) {
 	return p.client, nil
 }
 
-func (p *Provider) initAccount(ctx context.Context) (*Account, error) {
+func (p *Provider) initAccount() (*Account, error) {
 	if p.account == nil || len(p.account.Email) == 0 {
 		var err error
-		p.account, err = NewAccount(ctx, p.Email, p.KeyType)
+		p.account, err = NewAccount(p.Email)
 		if err != nil {
 			return nil, err
 		}
 	}
 
-	// Set the KeyType if not already defined in the account
-	if len(p.account.KeyType) == 0 {
-		p.account.KeyType = GetKeyType(ctx, p.KeyType)
-	}
-
 	return p.account, nil
 }
 
-func (p *Provider) register(ctx context.Context, client *lego.Client) (*registration.Resource, error) {
+func (p *Provider) register(ctx context.Context, client *lego.Client) (*acme.ExtendedAccount, error) {
 	logger := log.FromContext(ctx)
 
 	if p.EAB != nil {
@@ -356,12 +358,12 @@ func (p *Provider) register(ctx context.Context, client *lego.Client) (*registra
 
 		eabOptions := registration.RegisterEABOptions{TermsOfServiceAgreed: true, Kid: p.EAB.Kid, HmacEncoded: p.EAB.HmacEncoded}
 
-		return client.Registration.RegisterWithExternalAccountBinding(eabOptions)
+		return client.Registration.RegisterWithExternalAccountBinding(ctx, eabOptions)
 	}
 
 	logger.Info("Register...")
 
-	return client.Registration.Register(registration.RegisterOptions{TermsOfServiceAgreed: true})
+	return client.Registration.Register(ctx, registration.RegisterOptions{TermsOfServiceAgreed: true})
 }
 
 func (p *Provider) resolveDomains(ctx context.Context, domains []string, tlsStore string) {
@@ -571,12 +573,14 @@ func (p *Provider) resolveDefaultCertificate(ctx context.Context, domains []stri
 	}
 
 	request := certificate.ObtainRequest{
-		Domains:        domains,
-		Bundle:         true,
-		PreferredChain: p.PreferredChain,
+		Domains:          domains,
+		Bundle:           true,
+		PreferredChain:   p.PreferredChain,
+		EnableCommonName: true,
+		KeyType:          GetKeyType(ctx, p.KeyType),
 	}
 
-	cert, err := client.Certificate.Obtain(request)
+	cert, err := client.Certificate.Obtain(ctx, request)
 	if err != nil {
 		return nil, fmt.Errorf("unable to generate a certificate for the domains %v: %w", domains, err)
 	}
@@ -615,12 +619,14 @@ func (p *Provider) resolveCertificate(ctx context.Context, domain types.Domain, 
 	}
 
 	request := certificate.ObtainRequest{
-		Domains:        domains,
-		Bundle:         true,
-		PreferredChain: p.PreferredChain,
+		Domains:          domains,
+		Bundle:           true,
+		PreferredChain:   p.PreferredChain,
+		EnableCommonName: true,
+		KeyType:          GetKeyType(ctx, p.KeyType),
 	}
 
-	cert, err := client.Certificate.Obtain(request)
+	cert, err := client.Certificate.Obtain(ctx, request)
 	if err != nil {
 		return types.Domain{}, nil, fmt.Errorf("unable to generate a certificate for the domains %v: %w", uncheckedDomains, err)
 	}
@@ -808,7 +814,8 @@ func (p *Provider) renewCertificates(ctx context.Context, renewPeriod time.Durat
 		logger.Infof("Renewing certificate from LE : %+v", cert.Domain)
 
 		res := certificate.Resource{
-			Domain:      cert.Domain.Main,
+			ID:          cert.Domain.Main,
+			Domains:     cert.Domain.ToStrArray(),
 			PrivateKey:  cert.Key,
 			Certificate: cert.Certificate.Certificate,
 		}
@@ -818,7 +825,7 @@ func (p *Provider) renewCertificates(ctx context.Context, renewPeriod time.Durat
 			PreferredChain: p.PreferredChain,
 		}
 
-		renewedCert, err := client.Certificate.RenewWithOptions(res, opts)
+		renewedCert, err := client.Certificate.Renew(ctx, res, opts)
 		if err != nil {
 			logger.WithError(err).Errorf("Error renewing certificate from LE: %v", cert.Domain)
 			continue